pyproject.toml
[project]
name = "pydantic-ai-backend"
version = "0.2.0"
description = "File storage and sandbox backends for AI agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "VStorm" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0",
    "wcmatch>=8.0",
]

[project.optional-dependencies]
console = [
    "pydantic-ai-slim>=1.71.0",
]
daytona = [
    "daytona-sdk>=0.9.0",
]
docker = [
    "docker>=7.0",
    "pypdf>=5.1.0",
    "chardet>=5.2.0",
]
dev = [
    "pydantic-ai-slim>=1.71.0",
    "pytest>=9.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=7.0.0",
    "coverage>=7.0.0",
    "ruff>=0.8.0",
    "pyright>=1.1.400",
    "mypy>=1.0.0",
    "pre-commit>=4.0.0",
    "build>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/vstorm-co/pydantic-ai-backend"
Repository = "https://github.com/vstorm-co/pydantic-ai-backend"
Issues = "https://github.com/vstorm-co/pydantic-ai-backend/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pydantic_ai_backends"]

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # Pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "Q",      # flake8-quotes
    "RUF100", # unused noqa
    "C90",    # mccabe complexity
]

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.ruff.format]
quote-style = "double"

# Pytest configuration
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"
markers = [
    "docker: marks tests as requiring Docker (deselect with '-m \"not docker\"')",
]
addopts = "-m 'not docker'"

# Coverage configuration
[tool.coverage.run]
branch = true
source = ["src/pydantic_ai_backends"]
data_file = ".coverage/.coverage"

[tool.coverage.report]
fail_under = 100
skip_covered = true
show_missing = true
precision = 2
exclude_also = [
    "# pragma: no branch",
]
exclude_lines = [
    "pragma: no cover",
    "pragma: lax no cover",
    "if TYPE_CHECKING:",
    "@overload",
    "@abstractmethod",
    "raise NotImplementedError",
    "if __name__ == '__main__':",
    "except ImportError:",
    "pass",
    "\\.\\.\\.",
]

[tool.coverage.html]
directory = "htmlcov"

# Pyright configuration
[tool.pyright]
pythonVersion = "3.10"
typeCheckingMode = "basic"
include = ["src/pydantic_ai_backends"]
exclude = ["**/__pycache__", ".venv*", "tests"]
reportMissingTypeStubs = false
reportMissingImports = false
reportMissingModuleSource = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownVariableType = false
reportUnknownParameterType = false
reportOptionalMemberAccess = false

# MyPy configuration
[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# Codespell configuration
[tool.codespell]
skip = ".git,.venv*,uv.lock,htmlcov,.coverage"
